recombine ground_script.c
